backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
flask-migrate==4.0.7
pydantic==2.7.1
anthropic==0.28.0
python-dotenv==1.0.1
pytest==8.2.2
pytest-flask==1.3.0
